Add initial terrain

+++ src/chunk_volume_repository.hpp
@@ -10,12 +10,12 @@
 class ChunkVolumeRepository
 {
 public:
-    ChunkVolumeRepository(size_t bs) : border_size(bs) {}
+    ChunkVolumeRepository(int bs) : border_size(bs) {}
 
     template <typename F>
     void with(ChunkId, F);
 private:
-    const size_t border_size;
+    const int border_size;
 
     std::unordered_map<ChunkId, Volume<Voxel>> volumes;
     VolumeSampler volume_sampler;
